rename folder name / emotion styled 적용
